Update Lorenz63 backend to use float types and improve configuration handling
- Change data types from double to float in various components, including ConfigValue and state management.
- Update the Lorenz63 example to accept configuration parameters for initial state values.
- Modify CMakeLists.txt to link against new dependencies and ensure proper integration.
- Enhance tests to validate the new float-based configuration handling and state management.
- Improve overall consistency and performance across the Lorenz63 backend.

+++ applications/data_assimilation/ensemble/letkf.cpp
@@ -92,7 +92,7 @@ int main(int argc, char* argv[]) {
     try {
       ensemble_size = std::get<int>(config.Get("letkf.ensemble_size", 32));
       inflation_factor =
-          std::get<double>(config.Get("letkf.inflation_factor", 1.1));
+          std::get<float>(config.Get("letkf.inflation_factor", 1.1f));
       state_variables = std::get<std::vector<std::string>>(config.Get(
           "letkf.state_variables",
           std::vector<std::string>{"temperature", "pressure", "humidity"}));
